[C++] Split File joint-base : joint-dense is now in its own header file
