+++ CMakeLists.txt
@@ -87,6 +87,7 @@ SET(${PROJECT_NAME}_SPATIAL_HEADERS
 
 SET(${PROJECT_NAME}_MULTIBODY_JOINT_HEADERS
   multibody/joint/joint-base.hpp
+  multibody/joint/joint-dense.hpp
   multibody/joint/joint-revolute.hpp
   multibody/joint/joint-revolute-unaligned.hpp
   multibody/joint/joint-spherical.hpp
