+++ src/multibody/joint/joint-revolute-unaligned.hpp
@@ -19,6 +19,7 @@
 #define __se3_joint_revolute_unaligned_hpp__
 
 #include "pinocchio/multibody/joint/joint-base.hpp"
+#include "pinocchio/multibody/joint/joint-dense.hpp"
 #include "pinocchio/multibody/constraint.hpp"
 #include "pinocchio/spatial/inertia.hpp"
 #include "pinocchio/math/sincos.hpp"
