+++ src/multibody/joint/joint-revolute.hpp
@@ -22,6 +22,7 @@
 #include "pinocchio/math/sincos.hpp"
 #include "pinocchio/spatial/inertia.hpp"
 #include "pinocchio/multibody/joint/joint-base.hpp"
+#include "pinocchio/multibody/joint/joint-dense.hpp"
 
 namespace se3
 {
